splitting up tests that use config file at startup and tests that rely on command messages being sent to kafka
