+++ system-tests/compose/docker-compose-no-command.yml
@@ -1,0 +1,12 @@
+
+version: '2'
+
+services:
+  forwarder:
+    image: forwarder:latest
+    network_mode: "host"
+    volumes:
+      - ../config-files/forwarder_config_no_command.ini:/forwarder_config_no_command.ini
+      - ../logs/:/forwarder_logs/
+    environment:
+      CONFIG_FILE: "/forwarder_config_no_command.ini"
