+++ system-tests/config-files/forwarder_config_no_command.ini
@@ -1,0 +1,4 @@
+status-topic=//localhost:9092/TEST_forwarderStatus
+config-topic=//localhost:9092/TEST_forwarderConfig
+log-file=/forwarder_logs/forwarder_tests_no_command.log
+verbosity=7
